选手添加活动的时候编号必须存在且姓名相同才可以添加

+++ rate/rateserver/rate-service/src/main/java/org/sys/rate/service/admin/ParticipatesService.java
@@ -650,19 +650,26 @@ public class ParticipatesService {
 
     //该选手不在选手表的活动内，则在选手表添加记录，之后管理员导入根据编号更新补全信息。
     //该选手在选手表的活动内，但studentID不是他，则修改studentID，并在student表中删除该ID（老的）的记录
-    public void addActSelf(Integer studentID,Integer activityID,String code){
+    public boolean addActSelf(Integer studentID,Integer activityID,String code){
         // 首先检查选手表中是否存在在选手，通过code和activityID
         Participates participate = participatesMapper.getParticipateIDByCodeAndActivityID(code,activityID);
         if (participate == null){
-            // 不存在，则添加
-            participatesMapper.addPar(studentID,activityID,code);
+//            participatesMapper.addPar(studentID,activityID,code); // 不存在，则添加
+            return false; // 不存在，则不允许添加
         }else {
             // 存在，则检查studentID是否是他
             if (!Objects.equals(participate.getStudentID(), studentID)){
-                // 不是他，则修改studentID，并在student表中删除该ID（老的）的记录
-                participatesMapper.updatePar(studentID,participate.getID());
-                participatesMapper.deleteStudent(participate.getStudentID());
+                // 不是他，则核对姓名，如果姓名一样则修改studentID，并在student表中删除该ID（老的）的记录
+                String name = studentMapper.getNameByID(studentID);
+                if (name == null || !Objects.equals(name, participate.getName()))
+                    return false; // 姓名不一样，则不允许添加
+                else {
+                    // 姓名一样，则修改studentID，并在student表中删除该ID（老的）的记录
+                    participatesMapper.updatePar(studentID,participate.getID());
+                    participatesMapper.deleteStudent(participate.getStudentID());
+                }
             }
+            return true;
         }
     }
 }
